Feat/Automated IaC Deployment with Terraform, CircleCi, Docker, and Nginx (#32)
This pull request aims to improve the deployment process for the project by implementing automated and efficient system. It ensures that the deployment process is reliable and consistent, making it easier for developers to work on the project.

This pull request implements automated Infrastructure as Code (IaC) deployment using Terraform for Google Cloud Platform (GCP) and Cloudflare, CircleCi deployment workflows with jobs, Dockerfile and Docker Compose configurations, and a deployment scripts. Additionally, a reverse proxy with Nginx is configured for mainnet and testnet, and startup scripts are added for both boot node and peer nodes.

One of the major features of this pull request is the automated extraction of the bootnode ID and configuration of peernodes with that ID. This means that the deployment process is more streamlined and efficient, reducing the possibility of errors.

### Highlights

1. Configured Infrastructure as Code (IaC) with Terraform for GCP and Cloudflare.
2. Set up CircleCI deployment workflows with jobs.
3. Configured the Dockerfile and Docker Compose.
4. Added a deployment script.
5. Configured a reverse proxy with Nginx for the mainnet and testnet environments.
6. Configured a startup script for the boot node and peer nodes.
7. Automated the extraction of the bootnode ID and configured the peernodes with that ID.

### Missing Parts

1. Backup & Restore Automation
2. Jur Node Security

### Important Note
We need to turn off all the workflows for the mainnet once we have deployed and are ready to go live. This is important to prevent any changes to the mainnet by replacing the image. Instead, we will use manual runtime upgrades for the mainnet. This can be done by commenting out the CircleCI production workflow.